SAE3.02/requirements.txt
PyQt6 == 6.6.0
bcrypt == 4.1.1
mysql-connector-python == latest